Move storeges' ABCs to `storages/bases.py`
